requirements.txt
# API Framework
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1

# Database
sqlalchemy==2.0.35
pyodbc==5.2.0

# Data Processing
pandas==2.2.3
numpy==2.1.1

# Validation
pydantic==2.9.2

# CORS
python-multipart==0.0.9
